+++ app/lib/l10n/app_localizations_ur.dart
@@ -1,0 +1,35 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+import 'app_localizations.dart';
+
+class AppLocalizationsUr extends AppLocalizations {
+  AppLocalizationsUr([String locale = 'ur']) : super(locale);
+
+  @override
+  String get appName => 'انس لائف او ایس';
+
+  @override
+  String get startingTitle => 'آپ کی نجی جگہ تیار ہو رہی ہے';
+
+  @override
+  String get startingMessage => 'محفوظ آف لائن بنیاد کی جانچ ہو رہی ہے۔';
+
+  @override
+  String get foundationReadyTitle => 'بنیاد تیار ہے';
+
+  @override
+  String get foundationReadyMessage =>
+      'انس لائف او ایس اگلے منظور شدہ اسپرنٹ کے لیے تیار ہے۔ '
+      'ابھی کوئی ذاتی ڈیٹا یا پروڈکٹ فیچر فعال نہیں ہے۔';
+
+  @override
+  String get foundationErrorTitle => 'محفوظ بنیاد دستیاب نہیں';
+
+  @override
+  String get foundationErrorMessage =>
+      'مقامی انکرپٹڈ ڈیٹابیس انجن کی تصدیق نہیں ہو سکی۔ '
+      'کوئی ڈیٹا نہیں بنایا گیا۔ انسٹالیشن چیک کرکے دوبارہ کوشش کریں۔';
+
+  @override
+  String get retry => 'دوبارہ کوشش کریں';
+}